rename ArtworkFilterArtworkGrid2
