+++ src/v2/Apps/Artist/Routes/Overview/Components/ArtistArtworkFilter.tsx
@@ -138,7 +138,7 @@ export const ArtistArtworkFilterRefetchContainer = createRefetchContainer(
           width: $width
         ) {
           id
-          ...ArtworkFilterArtworkGrid2_filtered_artworks
+          ...ArtworkFilterArtworkGrid_filtered_artworks
         }
       }
     `,
