+++ src/v2/Apps/ArtistSeries/Components/ArtistSeriesArtworksFilter.tsx
@@ -101,7 +101,7 @@ export const ArtistSeriesArtworksFilterRefetchContainer = createRefetchContainer
           width: $width
         ) {
           id
-          ...ArtworkFilterArtworkGrid2_filtered_artworks
+          ...ArtworkFilterArtworkGrid_filtered_artworks
         }
       }
     `,
